web/src/requirements.txt
pyramid==1.10
pyramid-jinja2
requests==2.18
mysql-connector-python
python-dotenv
